refactor: tests use exporter for indices
