+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/opensearch/client/AbstractOpenSearchOperationIT.java
@@ -7,44 +7,49 @@
  */
 package io.camunda.operate.opensearch.client;
 
+import io.camunda.operate.conditions.DatabaseCondition;
 import io.camunda.operate.property.OperateProperties;
 import io.camunda.operate.schema.SchemaManager;
+import io.camunda.operate.schema.util.camunda.exporter.SchemaWithExporter;
 import io.camunda.operate.store.opensearch.client.sync.RichOpenSearchClient;
 import io.camunda.operate.util.OpensearchOperateAbstractIT;
 import io.camunda.operate.util.TestUtil;
 import java.util.function.Function;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.BeforeClass;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
+import org.springframework.test.context.TestPropertySource;
 
+@TestPropertySource(properties = DatabaseCondition.DATABASE_PROPERTY + "=opensearch")
 public abstract class AbstractOpenSearchOperationIT extends OpensearchOperateAbstractIT {
+  private static SchemaWithExporter schemaExporterHelper;
   @Autowired protected RichOpenSearchClient richOpenSearchClient;
-
   @Autowired protected SchemaManager schemaManager;
-
   @Autowired protected OperateProperties operateProperties;
-
   @Autowired protected OpensearchTestDataHelper opensearchTestDataHelper;
-
   protected String indexPrefix;
+
+  @BeforeClass
+  public static void beforeClass() {
+    schemaExporterHelper = new SchemaWithExporter("", false);
+  }
 
   @Before
   public void setUp() {
-    indexPrefix = "test-opensearch-operation-" + TestUtil.createRandomString(5);
-    operateProperties.getOpensearch().setIndexPrefix(indexPrefix);
-    schemaManager.createSchema();
+    schemaExporterHelper.createSchema();
   }
 
   @After
   public void cleanUp() {
-    schemaManager.deleteIndicesFor(indexPrefix + "*");
+    TestUtil.removeAllIndices(richOpenSearchClient.index(), richOpenSearchClient.template(), "*");
   }
 
-  public <R> R withThreadPoolTaskScheduler(Function<ThreadPoolTaskScheduler, R> f) {
+  public <R> R withThreadPoolTaskScheduler(final Function<ThreadPoolTaskScheduler, R> f) {
     final ThreadPoolTaskScheduler scheduler = new ThreadPoolTaskScheduler();
     scheduler.setPoolSize(5);
-    scheduler.setThreadNamePrefix(this.getClass().getSimpleName());
+    scheduler.setThreadNamePrefix(getClass().getSimpleName());
     scheduler.initialize();
 
     try {
