+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/util/OpensearchTestRuleProvider.java
@@ -16,6 +16,7 @@ import io.camunda.operate.exceptions.PersistenceException;
 import io.camunda.operate.property.OperateOpensearchProperties;
 import io.camunda.operate.property.OperateProperties;
 import io.camunda.operate.schema.SchemaManager;
+import io.camunda.operate.schema.util.camunda.exporter.SchemaWithExporter;
 import io.camunda.operate.store.opensearch.client.sync.RichOpenSearchClient;
 import io.camunda.operate.zeebe.ImportValueType;
 import io.camunda.operate.zeebeimport.RecordsReader;
@@ -112,7 +113,8 @@ public class OpensearchTestRuleProvider implements SearchTestRuleProvider {
       operateProperties.getOpensearch().setIndexPrefix(indexPrefix);
     }
     if (operateProperties.getOpensearch().isCreateSchema()) {
-      schemaManager.createSchema();
+      final var schemaExporterHelper = new SchemaWithExporter(indexPrefix, false);
+      schemaExporterHelper.createSchema();
       assertThat(areIndicesCreatedAfterChecks(indexPrefix, 5, 5 * 60 /*sec*/))
           .describedAs("Opensearch %s (min %d) indices are created", indexPrefix, 5)
           .isTrue();
